requirements.txt
adamp==0.3.0
art==6.1
colour==0.1.5
kornia==0.6.2
mmcv==2.1.0
mmdet==3.2.0
mmrotate==0.3.4
numpy==1.21.2
opencv_python==4.5.4.58
Shapely==2.0.2
Pillow==9.2.0
torch==1.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.62.2
wandb==0.10.15
